+++ crates/forge/src/lib.rs
@@ -35,7 +35,7 @@ mod warn;
 
 pub const CAIRO_EDITION: &str = "2024_07";
 
-const MINIMAL_RUST_VERSION: Version = Version::new(1, 80, 1);
+const MINIMAL_RUST_VERSION: Version = Version::new(1, 87, 0);
 const MINIMAL_SCARB_VERSION: Version = Version::new(2, 9, 1);
 const MINIMAL_RECOMMENDED_SCARB_VERSION: Version = Version::new(2, 9, 4);
 const MINIMAL_SCARB_VERSION_PREBUILT_PLUGIN: Version = Version::new(2, 10, 0);