+++ crates/snforge-scarb-plugin/src/attributes/available_gas.rs
@@ -6,7 +6,7 @@ use crate::{
     config_statement::extend_with_config_cheatcodes,
     types::{Number, ParseFromExpr},
 };
-use cairo_lang_macro::{quote, Diagnostic, Diagnostics, ProcMacroResult, Severity, TokenStream};
+use cairo_lang_macro::{Diagnostic, Diagnostics, ProcMacroResult, Severity, TokenStream, quote};
 use cairo_lang_parser::utils::SimpleParserDatabase;
 
 pub struct AvailableGasCollector;